refactor(template): colocate requirement leaf with composite
